% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simStateSpace-methods-statespacepb.R
\name{coef.pbstatespace}
\alias{coef.pbstatespace}
\title{Estimated Parameter Method for an Object of Class
\code{pbstatespace}}
\usage{
\method{coef}{pbstatespace}(object, ...)
}
\arguments{
\item{object}{Object of Class \code{pbstatespace}.}

\item{...}{additional arguments.}
}
\value{
Returns a vector of estimated parameters.
}
\description{
Estimated Parameter Method for an Object of Class
\code{pbstatespace}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
