% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simStateSpace-methods-statespacepb.R
\name{coef.statespacepb}
\alias{coef.statespacepb}
\title{Estimated Parameter Method for an Object of Class
\code{statespacepb}}
\usage{
\method{coef}{statespacepb}(object, ...)
}
\arguments{
\item{object}{Object of Class \code{statespacepb}.}

\item{...}{additional arguments.}
}
\value{
Returns a vector of estimated parameters.
}
\description{
Estimated Parameter Method for an Object of Class
\code{statespacepb}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
