% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{DVAR2SVAR1}
\alias{DVAR2SVAR1}
\title{Convert Parameters from the First Order
Discrete-Time Vector Autoregressive Model
to Structural Vector Autoregressive Model}
\usage{
DVAR2SVAR1(beta, psi_l, delta_t)
}
\arguments{
\item{beta}{Numeric matrix.
Transition matrix relating the values of the latent variables
at the previous to the current time point
(\eqn{\boldsymbol{\beta}}).}

\item{psi_l}{Numeric matrix.
Cholesky factorization (\code{t(chol(psi))})
of the covariance matrix
of the process noise
(\eqn{\boldsymbol{\Psi}}).}

\item{delta_t}{Numeric.
Time interval
(\eqn{\Delta_t}).}
}
\value{
Returns a list of structural vector autoregressive model parameters:
\itemize{
\item \code{betastar_0}: Numeric matrix.
Matrix of contemporaneous effects.
(\eqn{\boldsymbol{\beta}_{0}}).
\item \code{betastar_1}: Numeric matrix.
Transition matrix relating the values of the latent variables
from the previous time point to the current time point.
(\eqn{\boldsymbol{\beta}_{1}}).
\item \code{psistar_l}: Numeric matrix.
Cholesky factorization (\code{t(chol(psistar))})
of the process noise covariance matrix
\eqn{\boldsymbol{\Psi}^{\ast}}.
}
}
\description{
This function converts parameters from
the first order discrete-time vector autoregressive model
to structural vector autoregressive model.
}
\examples{
beta <- matrix(
  data = c(
    0.7,
    0.5,
    -0.1,
    0.0,
    0.6,
    0.4,
    0,
    0,
    0.5
  ),
  nrow = 3
)
psi <- 0.1 * diag(3)
psi_l <- t(chol(psi))
delta_t <- 0.10
DVAR2SVAR1(
  beta = beta,
  psi_l = psi_l,
  delta_t = delta_t
)

}
\seealso{
Other Simulation of State Space Models Data Functions: 
\code{\link{LinSDE2SSM}()},
\code{\link{SimSSMFixed}()},
\code{\link{SimSSMIVary}()},
\code{\link{SimSSMLinGrowth}()},
\code{\link{SimSSMLinGrowthIVary}()},
\code{\link{SimSSMLinSDEFixed}()},
\code{\link{SimSSMLinSDEIVary}()},
\code{\link{SimSSMOUFixed}()},
\code{\link{SimSSMOUIVary}()},
\code{\link{SimSSMVARFixed}()},
\code{\link{SimSSMVARIVary}()}
}
\concept{Simulation of State Space Models Data Functions}
\keyword{simStateSpace}
\keyword{svar}
\keyword{transformation}
\keyword{var}
