% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simStateSpace-methods-statespacepb.R
\name{extract.pbstatespace}
\alias{extract.pbstatespace}
\title{Extract Method for an Object of Class
\code{pbstatespace}}
\usage{
\method{extract}{pbstatespace}(object, what = NULL)
}
\arguments{
\item{object}{Object of Class \code{pbstatespace}.}

\item{what}{Character string.
What specific matrix to extract.
If \code{what = NULL},
extract all available matrices.}
}
\value{
Returns a list.
Each element of the list
is a list of bootstrap estimates
in matrix format.
}
\description{
Extract Method for an Object of Class
\code{pbstatespace}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
