% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simStateSpace-methods-statespacepb.R
\name{extract}
\alias{extract}
\title{Extract Generic Function}
\usage{
extract(object, what)
}
\arguments{
\item{object}{An object.}

\item{what}{Character string.}
}
\value{
A value determined by the specific method for the object's class.
}
\description{
A generic function for extracting elements from objects.
}
\keyword{methods}
