% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simStateSpace-methods-statespacepb.R
\name{extract.statespacepb}
\alias{extract.statespacepb}
\title{Extract Method for an Object of Class
\code{statespacepb}}
\usage{
\method{extract}{statespacepb}(object, what = NULL)
}
\arguments{
\item{object}{Object of Class \code{statespacepb}.}

\item{what}{Character string.
What specific matrix to extract.
If \code{what = NULL},
extract all available matrices.}
}
\value{
Returns a list.
Each element of the list
is a list of bootstrap estimates
in matrix format.
}
\description{
Extract Method for an Object of Class
\code{statespacepb}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
