% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LinSDE2SSM}
\alias{LinSDE2SSM}
\title{Convert Parameters from the Linear Stochastic Differential Equation Model
to State Space Model Parameterization}
\usage{
LinSDE2SSM(b, a, q, delta_t)
}
\arguments{
\item{b}{Numeric vector.
A term which is unobserved and constant over time
(\eqn{\mathbf{b}}).}

\item{a}{Numeric matrix.
The drift matrix
which represents the rate of change of the solution
in the absence of any random fluctuations
(\eqn{\mathbf{A}}).}

\item{q}{Numeric matrix.
The matrix of volatility
or randomness in the process
(\eqn{\mathbf{Q}}).}

\item{delta_t}{Numeric.
Time interval (\eqn{\delta_t}).}
}
\value{
Returns a list of state space parameters:
\itemize{
\item \code{alpha}: Numeric vector.
Vector of intercepts for the dynamic model
(\eqn{\boldsymbol{\alpha}}).
\item \code{beta}: Numeric matrix.
Transition matrix relating the values of the latent variables
at time \code{t_k - 1} to those at time \code{t_k}
(\eqn{\boldsymbol{\beta}}).
\item \code{psi}: Numeric matrix.
The process noise covariance matrix
(\eqn{\boldsymbol{\Psi}}).
}
}
\description{
This function converts parameters from
the linear stochastic differential equation model
to state space model parameterization.
}
\details{
The state space parameters
as a function of the linear stochastic differential equation model
parameters
are given by
\deqn{
      \boldsymbol{\beta}
      =
      \exp{
        \left(
          \mathbf{A}
          \Delta_{t}
        \right)
      }
  }

\deqn{
      \boldsymbol{\alpha}
      =
      \mathbf{A}^{-1}
      \left(
        \boldsymbol{\beta} - \mathbf{I}_{p}
      \right)
      \mathbf{b}
  }

\deqn{
      \mathrm{vec}
      \left(
        \boldsymbol{\Psi}
      \right)
      =
      \left\{
        \left[
          \left(
            \mathbf{A} \otimes \mathbf{I}_{p}
          \right)
          +
          \left(
            \mathbf{I}_{p} \otimes \mathbf{A}
          \right)
        \right]
        \left[
          \exp
          \left(
            \left[
              \left(
                \mathbf{A} \otimes \mathbf{I}_{p}
              \right)
              +
              \left(
                \mathbf{I}_{p} \otimes \mathbf{A}
              \right)
            \right]
            \Delta_{t}
        \right)
        -
        \mathbf{I}_{p \times p}
      \right]
      \mathrm{vec}
      \left(
        \mathbf{Q}
      \right)
    \right\}
  }
}
\examples{
p <- k <- 2
a <- matrix(
  data = c(
   -0.10,
   0.05,
   0.05,
   -0.10
 ),
 nrow = p
)
b <- c(0.317, 0.230)
q <- matrix(
  data = c(2.79, 0.06, 0.06, 3.27),
  nrow = p
)
delta_t <- 0.10

LinSDE2SSM(
  b = b,
  a = a,
  q = q,
  delta_t = delta_t
)

}
\seealso{
Other Simulation of State Space Models Data Functions: 
\code{\link{OU2SSM}()},
\code{\link{OU2SVAR}()},
\code{\link{OU2VAR}()},
\code{\link{SimSSM}()},
\code{\link{SimSSMFixed}()},
\code{\link{SimSSMIVary}()},
\code{\link{SimSSMLinGrowth}()},
\code{\link{SimSSMLinGrowthIVary}()},
\code{\link{SimSSMLinSDE}()},
\code{\link{SimSSMOU}()},
\code{\link{SimSSMOUFixed}()},
\code{\link{SimSSMOUIVary}()},
\code{\link{SimSSMVAR}()},
\code{\link{SimSSMVARFixed}()},
\code{\link{SimSSMVARIVary}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Simulation of State Space Models Data Functions}
\keyword{linsde}
\keyword{sim}
\keyword{simStateSpace}
