% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MuEta0}
\alias{MuEta0}
\title{Mean Vector of the Latent Variables
for a Stationary Process}
\usage{
MuEta0(beta, alpha)
}
\arguments{
\item{beta}{Numeric matrix.
Transition matrix relating the values of the latent variables
at the previous to the current time point
(\eqn{\boldsymbol{\beta}}).}

\item{alpha}{Numeric vector.
Vector of constant values for the dynamic model
(\eqn{\boldsymbol{\alpha}}).}
}
\description{
Mean Vector of the Latent Variables
for a Stationary Process
}
\details{
The mean vector of the latent variables
\eqn{\boldsymbol{\mu}_{\boldsymbol{\eta}, \boldsymbol{\eta}}}
is given by
\deqn{
    \boldsymbol{\mu}_{\boldsymbol{\eta}, \boldsymbol{\eta}}
    =
    \left(
      \mathbf{I} - \boldsymbol{\beta}
    \right)^{-1}
    \boldsymbol{\alpha}
  }
}
\examples{
p <- 3
beta <- 0.50 * diag(p)
alpha <- rep(x = 0.50, times = p)
MuEta0(beta = beta, alpha = alpha)

}
\seealso{
Other Simulation of State Space Models Data Functions: 
\code{\link{LinSDE2SSM}()},
\code{\link{SigmaEta0}()},
\code{\link{SimSSMFixed}()},
\code{\link{SimSSMIVary}()},
\code{\link{SimSSMLinGrowth}()},
\code{\link{SimSSMLinGrowthIVary}()},
\code{\link{SimSSMLinSDEFixed}()},
\code{\link{SimSSMLinSDEIVary}()},
\code{\link{SimSSMOUFixed}()},
\code{\link{SimSSMOUIVary}()},
\code{\link{SimSSMVARFixed}()},
\code{\link{SimSSMVARIVary}()}
}
\concept{Simulation of State Space Models Data Functions}
\keyword{sim}
\keyword{simStateSpace}
\keyword{ssm}
