% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{OU2SSM}
\alias{OU2SSM}
\title{Convert Parameters from the Ornstein–Uhlenbeck Model
to State Space Model Parameterization}
\usage{
OU2SSM(mu, phi, sigma_sqrt, delta_t)
}
\arguments{
\item{mu}{Numeric vector.
The long-term mean or equilibrium level
(\eqn{\boldsymbol{\mu}}).}

\item{phi}{Numeric matrix.
The rate of mean reversion,
determining how quickly the variable returns to its mean
(\eqn{\boldsymbol{\Phi}}).}

\item{sigma_sqrt}{Numeric matrix.
Cholesky decomposition of the matrix of volatility
or randomness in the process
(\eqn{\boldsymbol{\Sigma}}).}

\item{delta_t}{Numeric.
Time interval (\eqn{\delta_t}).}
}
\description{
This function converts parameters from the Ornstein–Uhlenbeck model
to state space model parameterization.
See details for more information.
}
\details{
The state space parameters
as a function of the  Ornstein–Uhlenbeck model parameters
are given by
\deqn{
      \boldsymbol{\beta}
      =
      \exp{
        \left(
          - \boldsymbol{\Phi}
          \Delta_{t}
        \right)
      }
  }

\deqn{
      \boldsymbol{\alpha}
      =
      - \boldsymbol{\Phi}^{-1}
      \left(
        \boldsymbol{\beta} - \mathbf{I}_{p}
      \right)
  }

\deqn{
      \mathrm{vec}
      \left(
        \boldsymbol{\Psi}
      \right)
      =
      \left\{
        \left[
          \left(
            - \boldsymbol{\Phi} \otimes \mathbf{I}_{p}
          \right)
          +
          \left(
            \mathbf{I}_{p} \otimes - \boldsymbol{\Phi}
          \right)
        \right]
        \left[
          \exp
          \left(
            \left[
              \left(
                - \boldsymbol{\Phi} \otimes \mathbf{I}_{p}
              \right)
              +
              \left(
                \mathbf{I}_{p} \otimes - \boldsymbol{\Phi}
              \right)
            \right]
            \Delta_{t}
        \right)
        -
        \mathbf{I}_{p \times p}
      \right]
      \mathrm{vec}
      \left(
        \boldsymbol{\Sigma}
      \right)
    \right\}
  }
}
\examples{
p <- k <- 2
mu <- c(5.76, 5.18)
phi <- matrix(data = c(0.10, -0.05, -0.05, 0.10), nrow = p)
sigma_sqrt <- chol(
  matrix(data = c(2.79, 0.06, 0.06, 3.27), nrow = p)
)
delta_t <- 0.10

OU2SSM(
  mu = mu,
  phi = phi,
  sigma_sqrt = sigma_sqrt,
  delta_t = delta_t
)

}
\seealso{
Other Simulation of State Space Models Data Functions: 
\code{\link{Sim2Matrix}()},
\code{\link{SimSSM0Fixed}()},
\code{\link{SimSSM0Vary}()},
\code{\link{SimSSM0}()},
\code{\link{SimSSMOUFixed}()},
\code{\link{SimSSMOUVary}()},
\code{\link{SimSSMOU}()},
\code{\link{SimSSMVARFixed}()},
\code{\link{SimSSMVARVary}()},
\code{\link{SimSSMVAR}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Simulation of State Space Models Data Functions}
\keyword{sim}
\keyword{simStateSpace}
