% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simStateSpace-ou-2-svar.R
\name{OU2SVAR}
\alias{OU2SVAR}
\title{Convert Parameters from the Ornstein–Uhlenbeck Model
to the Structural Vector Autoregressive Model}
\usage{
OU2SVAR(mu, phi, sigma, delta_t)
}
\arguments{
\item{mu}{Numeric vector.
The long-term mean or equilibrium level
(\eqn{\boldsymbol{\mu}}).}

\item{phi}{Numeric matrix.
The rate of mean reversion,
determining how quickly the variable returns to its mean
(\eqn{\boldsymbol{\Phi}}).}

\item{sigma}{Numeric matrix.
The matrix of volatility
or randomness in the process
(\eqn{\boldsymbol{\Sigma}}).}

\item{delta_t}{Numeric.
Discrete time interval (\eqn{\delta_t}).}
}
\description{
Convert Parameters from the Ornstein–Uhlenbeck Model
to the Structural Vector Autoregressive Model
}
\examples{
p <- k <- 2
mu <- c(5.76, 5.18)
phi <- matrix(
  data = c(0.10, -0.05, -0.05, 0.10),
  nrow = p
)
sigma <- matrix(
  data = c(2.79, 0.06, 0.06, 3.27),
  nrow = p
)
delta_t <- 1

OU2SVAR(
  mu = mu,
  phi = phi,
  sigma = sigma,
  delta_t = delta_t
)

}
\seealso{
Other Simulation of State Space Models Data Functions: 
\code{\link{OU2SSM}()},
\code{\link{OU2VAR}()},
\code{\link{SimSSMFixed}()},
\code{\link{SimSSMIVary}()},
\code{\link{SimSSMLinGrowthIVary}()},
\code{\link{SimSSMLinGrowth}()},
\code{\link{SimSSMOUFixed}()},
\code{\link{SimSSMOUIVary}()},
\code{\link{SimSSMOU}()},
\code{\link{SimSSMVARFixed}()},
\code{\link{SimSSMVARIVary}()},
\code{\link{SimSSMVAR}()},
\code{\link{SimSSM}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Simulation of State Space Models Data Functions}
\keyword{ou}
\keyword{sim}
\keyword{simStateSpace}
