% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simStateSpace-methods-statespacepb.R
\name{print.pbstatespace}
\alias{print.pbstatespace}
\title{Print Method for an Object of Class
\code{pbstatespace}}
\usage{
\method{print}{pbstatespace}(x, alpha = NULL, type = "pc", digits = 4, ...)
}
\arguments{
\item{x}{Object of Class \code{pbstatespace}.}

\item{alpha}{Numeric vector.
Significance level \eqn{\alpha}.
If \code{alpha = NULL},
use the argument \code{alpha} used in \code{x}.}

\item{type}{Charater string.
Confidence interval type, that is,
\code{type = "pc"} for percentile;
\code{type = "bc"} for bias corrected.}

\item{digits}{Digits to print.}

\item{...}{additional arguments.}
}
\value{
Prints a matrix of
estimates,
standard errors,
number of bootstrap replications,
and
confidence intervals.
}
\description{
Print Method for an Object of Class
\code{pbstatespace}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
