% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SigmaEta0}
\alias{SigmaEta0}
\title{Covariance Matrix of the Latent Variables
for a Stationary Process}
\usage{
SigmaEta0(beta, psi_l)
}
\arguments{
\item{beta}{Numeric matrix.
Transition matrix relating the values of the latent variables
at the previous to the current time point
(\eqn{\boldsymbol{\beta}}).}

\item{psi_l}{Numeric matrix.
Cholesky factorization (\code{t(chol(psi))})
of the covariance matrix
of the process noise
(\eqn{\boldsymbol{\Psi}}).}
}
\description{
Covariance Matrix of the Latent Variables
for a Stationary Process
}
\details{
The column vector of the elements
of the covariance matrix of the latent variables
\eqn{\boldsymbol{\Sigma}_{\boldsymbol{\eta}, \boldsymbol{\eta}}}
is given by
\deqn{
    \mathrm{Vec}
    \left(
      \boldsymbol{\Sigma}_{\boldsymbol{\eta}, \boldsymbol{\eta}}
    \right)
    =
    \left(
      \mathbf{I} - \boldsymbol{\beta} \otimes \boldsymbol{\beta}
    \right)^{-1}
    \mathrm{Vec}
    \left(
      \boldsymbol{\Psi}
    \right)
  }
}
\examples{
p <- 3
beta <- 0.50 * diag(p)
psi <- 0.001 * diag(p)
psi_l <- t(chol(psi))
SigmaEta0(beta = beta, psi_l = psi_l)

}
\seealso{
Other Simulation of State Space Models Data Functions: 
\code{\link{LinSDE2SSM}()},
\code{\link{MuEta0}()},
\code{\link{SimSSMFixed}()},
\code{\link{SimSSMIVary}()},
\code{\link{SimSSMLinGrowth}()},
\code{\link{SimSSMLinGrowthIVary}()},
\code{\link{SimSSMLinSDEFixed}()},
\code{\link{SimSSMLinSDEIVary}()},
\code{\link{SimSSMOUFixed}()},
\code{\link{SimSSMOUIVary}()},
\code{\link{SimSSMVARFixed}()},
\code{\link{SimSSMVARIVary}()}
}
\concept{Simulation of State Space Models Data Functions}
\keyword{sim}
\keyword{simStateSpace}
\keyword{ssm}
