% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simStateSpace-list-2-matrix.R
\name{Sim2Matrix}
\alias{Sim2Matrix}
\title{Simulation Output to Matrix}
\usage{
Sim2Matrix(x, eta = FALSE, long = TRUE)
}
\arguments{
\item{x}{R object.
Output of
\code{\link[=SimSSM]{SimSSM()}},
\code{\link[=SimSSMOU]{SimSSMOU()}},
\code{\link[=SimSSMVAR]{SimSSMVAR()}},
\code{\link[=SimSSMFixed]{SimSSMFixed()}},
\code{\link[=SimSSMOUFixed]{SimSSMOUFixed()}},
\code{\link[=SimSSMVARFixed]{SimSSMVARFixed()}},
\code{\link[=SimSSMIVary]{SimSSMIVary()}},
\code{\link[=SimSSMOUIVary]{SimSSMOUIVary()}}, or
\code{\link[=SimSSMVARIVary]{SimSSMVARIVary()}}.}

\item{eta}{Logical.
If \code{eta = TRUE}, include \code{eta}.
If \code{eta = FALSE}, exclude \code{eta}.}

\item{long}{Logical.
If \code{long = TRUE}, use long format.
If \code{long = FALSE}, use wide format.}
}
\value{
Returns a matrix of simulated data.
}
\description{
This function converts the output of
\code{\link[=SimSSM]{SimSSM()}},
\code{\link[=SimSSMOU]{SimSSMOU()}},
\code{\link[=SimSSMVAR]{SimSSMVAR()}},
\code{\link[=SimSSMFixed]{SimSSMFixed()}},
\code{\link[=SimSSMOUFixed]{SimSSMOUFixed()}},
\code{\link[=SimSSMVARFixed]{SimSSMVARFixed()}},
\code{\link[=SimSSMIVary]{SimSSMIVary()}},
\code{\link[=SimSSMOUIVary]{SimSSMOUIVary()}}, or
\code{\link[=SimSSMVARIVary]{SimSSMVARIVary()}}
to a matrix.
}
\examples{
# SimSSM Function -------------------------------------------------
# prepare parameters
set.seed(42)
k <- p <- 3
iden <- diag(k)
null_vec <- rep(x = 0, times = k)
mu0 <- null_vec
sigma0 <- iden
alpha <- null_vec
beta <- diag(x = 0.50, nrow = k)
psi <- iden
nu <- null_vec
lambda <- iden
theta <- diag(x = 0.50, nrow = k)
time <- 50
burn_in <- 0
gamma_y <- gamma_eta <- 0.10 * diag(k)
x <- matrix(
  data = rnorm(n = k * (time + burn_in)),
  ncol = k
)

# Type 0
ssm <- SimSSM(
  mu0 = mu0,
  sigma0 = sigma0,
  alpha = alpha,
  beta = beta,
  psi = psi,
  nu = nu,
  lambda = lambda,
  theta = theta,
  type = 0,
  time = time,
  burn_in = burn_in
)

mat <- Sim2Matrix(ssm, eta = TRUE)
head(mat)
mat <- Sim2Matrix(ssm, eta = FALSE)
head(mat)
mat <- Sim2Matrix(ssm, eta = TRUE, long = FALSE)
head(mat)
mat <- Sim2Matrix(ssm, eta = FALSE, long = FALSE)
head(mat)

# Type 1
ssm <- SimSSM(
  mu0 = mu0,
  sigma0 = sigma0,
  alpha = alpha,
  beta = beta,
  psi = psi,
  nu = nu,
  lambda = lambda,
  theta = theta,
  gamma_eta = gamma_eta,
  x = x,
  type = 1,
  time = time,
  burn_in = burn_in
)

mat <- Sim2Matrix(ssm, eta = TRUE)
head(mat)
mat <- Sim2Matrix(ssm, eta = FALSE)
head(mat)
mat <- Sim2Matrix(ssm, eta = TRUE, long = FALSE)
head(mat)
mat <- Sim2Matrix(ssm, eta = FALSE, long = FALSE)
head(mat)

# Type 2
ssm <- SimSSM(
  mu0 = mu0,
  sigma0 = sigma0,
  alpha = alpha,
  beta = beta,
  psi = psi,
  nu = nu,
  lambda = lambda,
  theta = theta,
  gamma_y = gamma_y,
  gamma_eta = gamma_eta,
  x = x,
  type = 2,
  time = time,
  burn_in = burn_in
)

mat <- Sim2Matrix(ssm, eta = TRUE)
head(mat)
mat <- Sim2Matrix(ssm, eta = FALSE)
head(mat)
mat <- Sim2Matrix(ssm, eta = TRUE, long = FALSE)
head(mat)
mat <- Sim2Matrix(ssm, eta = FALSE, long = FALSE)
head(mat)

}
\seealso{
Other Simulation of State Space Models Data Functions: 
\code{\link{OU2SSM}()},
\code{\link{SimSSMFixed}()},
\code{\link{SimSSMIVary}()},
\code{\link{SimSSMLinGrowthIVary}()},
\code{\link{SimSSMLinGrowth}()},
\code{\link{SimSSMOUFixed}()},
\code{\link{SimSSMOUIVary}()},
\code{\link{SimSSMOU}()},
\code{\link{SimSSMVARFixed}()},
\code{\link{SimSSMVARIVary}()},
\code{\link{SimSSMVAR}()},
\code{\link{SimSSM}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Simulation of State Space Models Data Functions}
\keyword{misc}
\keyword{simStateSpace}
