% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simStateSpace-sim-mvn.R
\name{SimIotaN}
\alias{SimIotaN}
\title{Simulate Intercept Vectors
in a Continuous-Time Vector Autoregressive Model
from the Multivariate Normal Distribution}
\usage{
SimIotaN(n, iota, vcov_iota_l)
}
\arguments{
\item{n}{Positive integer.
Number of replications.}

\item{iota}{Numeric vector.
Intercept (\eqn{\boldsymbol{\iota}}).}

\item{vcov_iota_l}{Numeric matrix.
Cholesky factorization (\code{t(chol(vcov_iota))})
of the sampling variance-covariance matrix of
\eqn{\boldsymbol{\iota}}.}
}
\value{
Returns a list of random intercept vectors.
}
\description{
This function simulates random intercept vectors
in a continuous-time vector autoregressive model
from the multivariate normal distribution.
}
\examples{
n <- 10
iota <- c(0, 0, 0)
vcov_iota_l <- t(chol(0.001 * diag(3)))
SimIotaN(n = n, iota = iota, vcov_iota_l = vcov_iota_l)

}
\seealso{
Other Simulation of State Space Models Data Functions:
\code{\link[=LinSDE2SSM]{LinSDE2SSM()}},
\code{\link[=LinSDECovEta]{LinSDECovEta()}},
\code{\link[=LinSDECovY]{LinSDECovY()}},
\code{\link[=LinSDEInterceptEta]{LinSDEInterceptEta()}},
\code{\link[=LinSDEInterceptY]{LinSDEInterceptY()}},
\code{\link[=LinSDEMeanEta]{LinSDEMeanEta()}},
\code{\link[=LinSDEMeanY]{LinSDEMeanY()}},
\code{\link[=ProjectToHurwitz]{ProjectToHurwitz()}},
\code{\link[=ProjectToStability]{ProjectToStability()}},
\code{\link[=SSMCovEta]{SSMCovEta()}},
\code{\link[=SSMCovY]{SSMCovY()}},
\code{\link[=SSMInterceptEta]{SSMInterceptEta()}},
\code{\link[=SSMInterceptY]{SSMInterceptY()}},
\code{\link[=SSMMeanEta]{SSMMeanEta()}},
\code{\link[=SSMMeanY]{SSMMeanY()}},
\code{\link[=SimAlphaN]{SimAlphaN()}},
\code{\link[=SimBetaN]{SimBetaN()}},
\code{\link[=SimBetaN2]{SimBetaN2()}},
\code{\link[=SimBetaNCovariate]{SimBetaNCovariate()}},
\code{\link[=SimCovDiagN]{SimCovDiagN()}},
\code{\link[=SimCovN]{SimCovN()}},
\code{\link[=SimMVN]{SimMVN()}},
\code{\link[=SimMuN]{SimMuN()}},
\code{\link[=SimNuN]{SimNuN()}},
\code{\link[=SimPhiN]{SimPhiN()}},
\code{\link[=SimPhiN2]{SimPhiN2()}},
\code{\link[=SimPhiNCovariate]{SimPhiNCovariate()}},
\code{\link[=SimSSMFixed]{SimSSMFixed()}},
\code{\link[=SimSSMIVary]{SimSSMIVary()}},
\code{\link[=SimSSMLinGrowth]{SimSSMLinGrowth()}},
\code{\link[=SimSSMLinGrowthIVary]{SimSSMLinGrowthIVary()}},
\code{\link[=SimSSMLinSDEFixed]{SimSSMLinSDEFixed()}},
\code{\link[=SimSSMLinSDEIVary]{SimSSMLinSDEIVary()}},
\code{\link[=SimSSMOUFixed]{SimSSMOUFixed()}},
\code{\link[=SimSSMOUIVary]{SimSSMOUIVary()}},
\code{\link[=SimSSMVARFixed]{SimSSMVARFixed()}},
\code{\link[=SimSSMVARIVary]{SimSSMVARIVary()}},
\code{\link[=SpectralRadius]{SpectralRadius()}},
\code{\link[=TestPhi]{TestPhi()}},
\code{\link[=TestPhiHurwitz]{TestPhiHurwitz()}},
\code{\link[=TestStability]{TestStability()}},
\code{\link[=TestStationarity]{TestStationarity()}}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Simulation of State Space Models Data Functions}
\keyword{simStateSpace}
\keyword{ssm}
