% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simStateSpace-sim-mvn.R
\name{SimMVN}
\alias{SimMVN}
\title{Simulate Vectors
from the Multivariate Normal Distribution}
\usage{
SimMVN(n, mu, sigma_l)
}
\arguments{
\item{n}{Positive integer.
Number of replications.}

\item{mu}{Numeric vector.
Mean vector (\eqn{\boldsymbol{\nu}}).}

\item{sigma_l}{Numeric matrix.
Cholesky factorization (\code{t(chol(sigma))})
of the variance-covariance matrix
\eqn{\boldsymbol{\Sigma}}.}
}
\value{
Returns a list of random vectors.
}
\description{
This function simulates random vectors
from the multivariate normal distribution.
}
\examples{
n <- 10
mu <- c(0, 0, 0)
sigma_l <- t(chol(0.001 * diag(3)))
SimMVN(n = n, mu = mu, sigma_l = sigma_l)

}
\seealso{
Other Simulation of State Space Models Data Functions:
\code{\link{LinSDE2SSM}()},
\code{\link{LinSDECovEta}()},
\code{\link{LinSDECovY}()},
\code{\link{LinSDEInterceptEta}()},
\code{\link{LinSDEInterceptY}()},
\code{\link{LinSDEMeanEta}()},
\code{\link{LinSDEMeanY}()},
\code{\link{ProjectToHurwitz}()},
\code{\link{ProjectToStability}()},
\code{\link{SSMCovEta}()},
\code{\link{SSMCovY}()},
\code{\link{SSMInterceptEta}()},
\code{\link{SSMInterceptY}()},
\code{\link{SSMMeanEta}()},
\code{\link{SSMMeanY}()},
\code{\link{SimAlphaN}()},
\code{\link{SimBetaN}()},
\code{\link{SimBetaN2}()},
\code{\link{SimBetaNCovariate}()},
\code{\link{SimCovDiagN}()},
\code{\link{SimCovN}()},
\code{\link{SimIotaN}()},
\code{\link{SimMuN}()},
\code{\link{SimNuN}()},
\code{\link{SimPhiN}()},
\code{\link{SimPhiN2}()},
\code{\link{SimPhiNCovariate}()},
\code{\link{SimSSMFixed}()},
\code{\link{SimSSMIVary}()},
\code{\link{SimSSMLinGrowth}()},
\code{\link{SimSSMLinGrowthIVary}()},
\code{\link{SimSSMLinSDEFixed}()},
\code{\link{SimSSMLinSDEIVary}()},
\code{\link{SimSSMOUFixed}()},
\code{\link{SimSSMOUIVary}()},
\code{\link{SimSSMVARFixed}()},
\code{\link{SimSSMVARIVary}()},
\code{\link{SpectralRadius}()},
\code{\link{TestPhi}()},
\code{\link{TestPhiHurwitz}()},
\code{\link{TestStability}()},
\code{\link{TestStationarity}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Simulation of State Space Models Data Functions}
\keyword{simStateSpace}
\keyword{ssm}
