% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simStateSpace-package.R
\docType{package}
\name{simStateSpace-package}
\alias{simStateSpace-package}
\alias{_PACKAGE}
\title{simStateSpace: Simulate Data from State Space Models}
\description{
Offers an efficient and user-friendly framework for simulating data in state space models where n subjects/units is greater than one which is common in social and behavioral sciences.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jeksterslab/simStateSpace}
  \item \url{https://jeksterslab.github.io/simStateSpace/}
  \item Report bugs at \url{https://github.com/jeksterslab/simStateSpace/issues}
}

}
\author{
\strong{Maintainer}: Ivan Jacob Agaloos Pesigan \email{r.jeksterslab@gmail.com} (\href{https://orcid.org/0000-0003-4818-8420}{ORCID}) [copyright holder]

}
\keyword{internal}
