% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SSMCov}
\alias{SSMCov}
\title{State Covariance Matrix for the
State Space Model}
\usage{
SSMCov(beta, psi)
}
\arguments{
\item{beta}{Numeric matrix.
The transition matrix (\eqn{\boldsymbol{\beta}}).}

\item{psi}{Numeric matrix.
The covariance matrix
of the process noise
(\eqn{\boldsymbol{\Psi}}).}
}
\description{
This function calculates the state covariance matrix
for the state space model
given by
\deqn{
  \mathrm{vec}
  \left(
    \mathrm{Cov} \left( \boldsymbol{\eta} \right)
  \right)
  =
  \left(
    \mathbf{I} - \boldsymbol{\beta} \otimes \boldsymbol{\beta}
  \right)^{-1}
  \mathrm{vec} \left( \boldsymbol{\Psi} \right) .
}
}
\examples{
beta <- 0.50 * diag(3)
psi <- 0.001 * diag(3)
SSMCov(beta = beta, psi = psi)

}
\seealso{
Other Simulation of State Space Models Data Functions: 
\code{\link{LinSDE2SSM}()},
\code{\link{LinSDECov}()},
\code{\link{LinSDEMean}()},
\code{\link{SSMMean}()},
\code{\link{SimAlphaN}()},
\code{\link{SimBetaN}()},
\code{\link{SimCovDiagN}()},
\code{\link{SimCovN}()},
\code{\link{SimIotaN}()},
\code{\link{SimPhiN}()},
\code{\link{SimSSMFixed}()},
\code{\link{SimSSMIVary}()},
\code{\link{SimSSMLinGrowth}()},
\code{\link{SimSSMLinGrowthIVary}()},
\code{\link{SimSSMLinSDEFixed}()},
\code{\link{SimSSMLinSDEIVary}()},
\code{\link{SimSSMOUFixed}()},
\code{\link{SimSSMOUIVary}()},
\code{\link{SimSSMVARFixed}()},
\code{\link{SimSSMVARIVary}()},
\code{\link{TestPhi}()},
\code{\link{TestStability}()},
\code{\link{TestStationarity}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Simulation of State Space Models Data Functions}
\keyword{simStateSpace}
\keyword{ssm}
