% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simStateSpace-ssm-cov-y.R
\name{SSMCovY}
\alias{SSMCovY}
\title{Steady-State Covariance Matrix for the
Observed Variables in the
State Space Model}
\usage{
SSMCovY(lambda, theta, cov_eta)
}
\arguments{
\item{lambda}{Numeric matrix.
Factor loading matrix linking the latent variables
to the observed variables
(\eqn{\boldsymbol{\Lambda}}).}

\item{theta}{Numeric matrix.
The covariance matrix of
the measurement error
(\eqn{\boldsymbol{\Theta}}).}

\item{cov_eta}{Numeric matrix.
The steady-state covariance matrix
for the latent variables
in the state space model}
}
\description{
The steady-state covariance matrix
for the observed variables
in the state space model
\eqn{\mathrm{Cov} \left( \mathbf{y} \right)}
is given by
\deqn{
  \mathrm{Cov} \left( \mathbf{y} \right)
  =
  \boldsymbol{\Lambda}
  \mathrm{Cov} \left( \boldsymbol{\eta} \right)
  \boldsymbol{\Lambda}^{\prime}
  +
  \boldsymbol{\Theta}
}
where
\eqn{\boldsymbol{\Lambda}}
is the matrix of factor loadings,
\eqn{\boldsymbol{\Theta}}
is the covariance matrix of
the measurement error,
and
\eqn{\mathrm{Cov} \left( \boldsymbol{\eta} \right)}
is the steady-state covariance matrix
for the latent variables.
}
\examples{
beta <- matrix(
  data = c(
    0.7, 0.5, -0.1,
    0.0, 0.6, 0.4,
    0.0, 0.0, 0.5
  ),
  nrow = 3
)
psi <- 0.1 * diag(3)
lambda <- diag(3)
theta <- diag(3)
cov_eta <- SSMCovEta(
  beta = beta,
  psi = psi
)
SSMCovY(
  lambda = lambda,
  theta = theta,
  cov_eta = cov_eta
)

}
\seealso{
Other Simulation of State Space Models Data Functions:
\code{\link[=LinSDE2SSM]{LinSDE2SSM()}},
\code{\link[=LinSDECovEta]{LinSDECovEta()}},
\code{\link[=LinSDECovY]{LinSDECovY()}},
\code{\link[=LinSDEInterceptEta]{LinSDEInterceptEta()}},
\code{\link[=LinSDEInterceptY]{LinSDEInterceptY()}},
\code{\link[=LinSDEMeanEta]{LinSDEMeanEta()}},
\code{\link[=LinSDEMeanY]{LinSDEMeanY()}},
\code{\link[=ProjectToHurwitz]{ProjectToHurwitz()}},
\code{\link[=ProjectToStability]{ProjectToStability()}},
\code{\link[=SSMCovEta]{SSMCovEta()}},
\code{\link[=SSMInterceptEta]{SSMInterceptEta()}},
\code{\link[=SSMInterceptY]{SSMInterceptY()}},
\code{\link[=SSMMeanEta]{SSMMeanEta()}},
\code{\link[=SSMMeanY]{SSMMeanY()}},
\code{\link[=SimAlphaN]{SimAlphaN()}},
\code{\link[=SimBetaN]{SimBetaN()}},
\code{\link[=SimBetaN2]{SimBetaN2()}},
\code{\link[=SimBetaNCovariate]{SimBetaNCovariate()}},
\code{\link[=SimCovDiagN]{SimCovDiagN()}},
\code{\link[=SimCovN]{SimCovN()}},
\code{\link[=SimIotaN]{SimIotaN()}},
\code{\link[=SimMVN]{SimMVN()}},
\code{\link[=SimMuN]{SimMuN()}},
\code{\link[=SimNuN]{SimNuN()}},
\code{\link[=SimPhiN]{SimPhiN()}},
\code{\link[=SimPhiN2]{SimPhiN2()}},
\code{\link[=SimPhiNCovariate]{SimPhiNCovariate()}},
\code{\link[=SimSSMFixed]{SimSSMFixed()}},
\code{\link[=SimSSMIVary]{SimSSMIVary()}},
\code{\link[=SimSSMLinGrowth]{SimSSMLinGrowth()}},
\code{\link[=SimSSMLinGrowthIVary]{SimSSMLinGrowthIVary()}},
\code{\link[=SimSSMLinSDEFixed]{SimSSMLinSDEFixed()}},
\code{\link[=SimSSMLinSDEIVary]{SimSSMLinSDEIVary()}},
\code{\link[=SimSSMOUFixed]{SimSSMOUFixed()}},
\code{\link[=SimSSMOUIVary]{SimSSMOUIVary()}},
\code{\link[=SimSSMVARFixed]{SimSSMVARFixed()}},
\code{\link[=SimSSMVARIVary]{SimSSMVARIVary()}},
\code{\link[=SpectralRadius]{SpectralRadius()}},
\code{\link[=TestPhi]{TestPhi()}},
\code{\link[=TestPhiHurwitz]{TestPhiHurwitz()}},
\code{\link[=TestStability]{TestStability()}},
\code{\link[=TestStationarity]{TestStationarity()}}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Simulation of State Space Models Data Functions}
\keyword{simStateSpace}
\keyword{ssm}
