% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simStateSpace-ssm-mean-y.R
\name{SSMMeanY}
\alias{SSMMeanY}
\title{Steady-State Mean Vector for the
Observed Variables in the
State Space Model}
\usage{
SSMMeanY(nu, lambda, mean_eta)
}
\arguments{
\item{nu}{Numeric vector.
Vector of intercept values for the measurement model
(\eqn{\boldsymbol{\nu}}).}

\item{lambda}{Numeric matrix.
Factor loading matrix linking the latent variables
to the observed variables
(\eqn{\boldsymbol{\Lambda}}).}

\item{mean_eta}{Numeric vector.
Steady-state mean vector
of the latent variables
\eqn{\mathrm{Mean} \left( \boldsymbol{\eta} \right)}.}
}
\description{
The steady-state mean vector
for the observed variables
in the state space model
\eqn{\mathrm{Mean} \left( \mathbf{y} \right)}
is given by
\deqn{
  \mathrm{Mean} \left( \mathbf{y} \right)
  =
  \boldsymbol{\nu}
  +
  \boldsymbol{\Lambda}
  \mathrm{Mean} \left( \boldsymbol{\eta} \right)
}
where
\eqn{\boldsymbol{\nu}}
is the vector of intercept values
for the measurement model,
\eqn{\boldsymbol{\Lambda}}
is the matrix of factor loadings,
and
\eqn{\mathrm{Mean} \left( \boldsymbol{\eta} \right)}
is the steady-state mean vector
for the latent variables.
}
\examples{
beta <- matrix(
  data = c(
    0.7, 0.5, -0.1,
    0.0, 0.6, 0.4,
    0.0, 0.0, 0.5
  ),
  nrow = 3
)
alpha <- rep(x = 1, times = 3)
lambda <- diag(3)
nu <- rep(x = 1, times = 3)
mean_eta <- SSMMeanEta(
  beta = beta,
  alpha = alpha
)
SSMMeanY(
  nu = nu,
  lambda = lambda,
  mean_eta = mean_eta
)

}
\seealso{
Other Simulation of State Space Models Data Functions:
\code{\link[=LinSDE2SSM]{LinSDE2SSM()}},
\code{\link[=LinSDECovEta]{LinSDECovEta()}},
\code{\link[=LinSDECovY]{LinSDECovY()}},
\code{\link[=LinSDEInterceptEta]{LinSDEInterceptEta()}},
\code{\link[=LinSDEInterceptY]{LinSDEInterceptY()}},
\code{\link[=LinSDEMeanEta]{LinSDEMeanEta()}},
\code{\link[=LinSDEMeanY]{LinSDEMeanY()}},
\code{\link[=ProjectToHurwitz]{ProjectToHurwitz()}},
\code{\link[=ProjectToStability]{ProjectToStability()}},
\code{\link[=SSMCovEta]{SSMCovEta()}},
\code{\link[=SSMCovY]{SSMCovY()}},
\code{\link[=SSMInterceptEta]{SSMInterceptEta()}},
\code{\link[=SSMInterceptY]{SSMInterceptY()}},
\code{\link[=SSMMeanEta]{SSMMeanEta()}},
\code{\link[=SimAlphaN]{SimAlphaN()}},
\code{\link[=SimBetaN]{SimBetaN()}},
\code{\link[=SimBetaN2]{SimBetaN2()}},
\code{\link[=SimBetaNCovariate]{SimBetaNCovariate()}},
\code{\link[=SimCovDiagN]{SimCovDiagN()}},
\code{\link[=SimCovN]{SimCovN()}},
\code{\link[=SimIotaN]{SimIotaN()}},
\code{\link[=SimMVN]{SimMVN()}},
\code{\link[=SimMuN]{SimMuN()}},
\code{\link[=SimNuN]{SimNuN()}},
\code{\link[=SimPhiN]{SimPhiN()}},
\code{\link[=SimPhiN2]{SimPhiN2()}},
\code{\link[=SimPhiNCovariate]{SimPhiNCovariate()}},
\code{\link[=SimSSMFixed]{SimSSMFixed()}},
\code{\link[=SimSSMIVary]{SimSSMIVary()}},
\code{\link[=SimSSMLinGrowth]{SimSSMLinGrowth()}},
\code{\link[=SimSSMLinGrowthIVary]{SimSSMLinGrowthIVary()}},
\code{\link[=SimSSMLinSDEFixed]{SimSSMLinSDEFixed()}},
\code{\link[=SimSSMLinSDEIVary]{SimSSMLinSDEIVary()}},
\code{\link[=SimSSMOUFixed]{SimSSMOUFixed()}},
\code{\link[=SimSSMOUIVary]{SimSSMOUIVary()}},
\code{\link[=SimSSMVARFixed]{SimSSMVARFixed()}},
\code{\link[=SimSSMVARIVary]{SimSSMVARIVary()}},
\code{\link[=SpectralRadius]{SpectralRadius()}},
\code{\link[=TestPhi]{TestPhi()}},
\code{\link[=TestPhiHurwitz]{TestPhiHurwitz()}},
\code{\link[=TestStability]{TestStability()}},
\code{\link[=TestStationarity]{TestStationarity()}}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Simulation of State Space Models Data Functions}
\keyword{simStateSpace}
\keyword{ssm}
