% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{TestStability}
\alias{TestStability}
\title{Test Stability}
\usage{
TestStability(x, margin = 0)
}
\arguments{
\item{x}{Numeric matrix.}

\item{margin}{Numeric scalar specifying the stability threshold
for the real part of the eigenvalues.
The default \code{0.0} corresponds to the imaginary axis;
values less than \code{0.0} enforce a stricter stability margin.}
}
\description{
The function computes the eigenvalues of the input matrix \code{x}.
It checks if the real part of all eigenvalues is negative.
If all eigenvalues have negative real parts,
the system is considered stable.
}
\examples{
x <- matrix(
  data = c(
    -0.357, 0.771, -0.450,
    0.0, -0.511, 0.729,
    0, 0, -0.693
  ),
  nrow = 3
)
TestStability(x)

}
\seealso{
Other Simulation of State Space Models Data Functions:
\code{\link[=LinSDE2SSM]{LinSDE2SSM()}},
\code{\link[=LinSDECovEta]{LinSDECovEta()}},
\code{\link[=LinSDECovY]{LinSDECovY()}},
\code{\link[=LinSDEInterceptEta]{LinSDEInterceptEta()}},
\code{\link[=LinSDEInterceptY]{LinSDEInterceptY()}},
\code{\link[=LinSDEMeanEta]{LinSDEMeanEta()}},
\code{\link[=LinSDEMeanY]{LinSDEMeanY()}},
\code{\link[=ProjectToHurwitz]{ProjectToHurwitz()}},
\code{\link[=ProjectToStability]{ProjectToStability()}},
\code{\link[=SSMCovEta]{SSMCovEta()}},
\code{\link[=SSMCovY]{SSMCovY()}},
\code{\link[=SSMInterceptEta]{SSMInterceptEta()}},
\code{\link[=SSMInterceptY]{SSMInterceptY()}},
\code{\link[=SSMMeanEta]{SSMMeanEta()}},
\code{\link[=SSMMeanY]{SSMMeanY()}},
\code{\link[=SimAlphaN]{SimAlphaN()}},
\code{\link[=SimBetaN]{SimBetaN()}},
\code{\link[=SimBetaN2]{SimBetaN2()}},
\code{\link[=SimBetaNCovariate]{SimBetaNCovariate()}},
\code{\link[=SimCovDiagN]{SimCovDiagN()}},
\code{\link[=SimCovN]{SimCovN()}},
\code{\link[=SimIotaN]{SimIotaN()}},
\code{\link[=SimMVN]{SimMVN()}},
\code{\link[=SimMuN]{SimMuN()}},
\code{\link[=SimNuN]{SimNuN()}},
\code{\link[=SimPhiN]{SimPhiN()}},
\code{\link[=SimPhiN2]{SimPhiN2()}},
\code{\link[=SimPhiNCovariate]{SimPhiNCovariate()}},
\code{\link[=SimSSMFixed]{SimSSMFixed()}},
\code{\link[=SimSSMIVary]{SimSSMIVary()}},
\code{\link[=SimSSMLinGrowth]{SimSSMLinGrowth()}},
\code{\link[=SimSSMLinGrowthIVary]{SimSSMLinGrowthIVary()}},
\code{\link[=SimSSMLinSDEFixed]{SimSSMLinSDEFixed()}},
\code{\link[=SimSSMLinSDEIVary]{SimSSMLinSDEIVary()}},
\code{\link[=SimSSMOUFixed]{SimSSMOUFixed()}},
\code{\link[=SimSSMOUIVary]{SimSSMOUIVary()}},
\code{\link[=SimSSMVARFixed]{SimSSMVARFixed()}},
\code{\link[=SimSSMVARIVary]{SimSSMVARIVary()}},
\code{\link[=SpectralRadius]{SpectralRadius()}},
\code{\link[=TestPhi]{TestPhi()}},
\code{\link[=TestPhiHurwitz]{TestPhiHurwitz()}},
\code{\link[=TestStationarity]{TestStationarity()}}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Simulation of State Space Models Data Functions}
\keyword{linsde}
\keyword{simStateSpace}
\keyword{test}
