% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{TestStationarity}
\alias{TestStationarity}
\title{Test Stationarity}
\usage{
TestStationarity(x, margin = 1)
}
\arguments{
\item{x}{Numeric matrix.}

\item{margin}{Numeric scalar specifying the stationarity threshold.
Values less than 1 indicate stricter stationarity criteria.}
}
\description{
The function computes the eigenvalues of the input matrix \code{x}.
It checks if all eigenvalues have moduli less than 1.
If all eigenvalues have moduli less than 1,
the system is considered stationary.
}
\examples{
x <- matrix(
  data = c(0.5, 0.3, 0.2, 0.4),
  nrow = 2
)
TestStationarity(x)

x <- matrix(
  data = c(0.9, -0.5, 0.8, 0.7),
  nrow = 2
)
TestStationarity(x)

}
\seealso{
Other Simulation of State Space Models Data Functions:
\code{\link[=LinSDE2SSM]{LinSDE2SSM()}},
\code{\link[=LinSDECovEta]{LinSDECovEta()}},
\code{\link[=LinSDECovY]{LinSDECovY()}},
\code{\link[=LinSDEInterceptEta]{LinSDEInterceptEta()}},
\code{\link[=LinSDEInterceptY]{LinSDEInterceptY()}},
\code{\link[=LinSDEMeanEta]{LinSDEMeanEta()}},
\code{\link[=LinSDEMeanY]{LinSDEMeanY()}},
\code{\link[=ProjectToHurwitz]{ProjectToHurwitz()}},
\code{\link[=ProjectToStability]{ProjectToStability()}},
\code{\link[=SSMCovEta]{SSMCovEta()}},
\code{\link[=SSMCovY]{SSMCovY()}},
\code{\link[=SSMInterceptEta]{SSMInterceptEta()}},
\code{\link[=SSMInterceptY]{SSMInterceptY()}},
\code{\link[=SSMMeanEta]{SSMMeanEta()}},
\code{\link[=SSMMeanY]{SSMMeanY()}},
\code{\link[=SimAlphaN]{SimAlphaN()}},
\code{\link[=SimBetaN]{SimBetaN()}},
\code{\link[=SimBetaN2]{SimBetaN2()}},
\code{\link[=SimBetaNCovariate]{SimBetaNCovariate()}},
\code{\link[=SimCovDiagN]{SimCovDiagN()}},
\code{\link[=SimCovN]{SimCovN()}},
\code{\link[=SimIotaN]{SimIotaN()}},
\code{\link[=SimMVN]{SimMVN()}},
\code{\link[=SimMuN]{SimMuN()}},
\code{\link[=SimNuN]{SimNuN()}},
\code{\link[=SimPhiN]{SimPhiN()}},
\code{\link[=SimPhiN2]{SimPhiN2()}},
\code{\link[=SimPhiNCovariate]{SimPhiNCovariate()}},
\code{\link[=SimSSMFixed]{SimSSMFixed()}},
\code{\link[=SimSSMIVary]{SimSSMIVary()}},
\code{\link[=SimSSMLinGrowth]{SimSSMLinGrowth()}},
\code{\link[=SimSSMLinGrowthIVary]{SimSSMLinGrowthIVary()}},
\code{\link[=SimSSMLinSDEFixed]{SimSSMLinSDEFixed()}},
\code{\link[=SimSSMLinSDEIVary]{SimSSMLinSDEIVary()}},
\code{\link[=SimSSMOUFixed]{SimSSMOUFixed()}},
\code{\link[=SimSSMOUIVary]{SimSSMOUIVary()}},
\code{\link[=SimSSMVARFixed]{SimSSMVARFixed()}},
\code{\link[=SimSSMVARIVary]{SimSSMVARIVary()}},
\code{\link[=SpectralRadius]{SpectralRadius()}},
\code{\link[=TestPhi]{TestPhi()}},
\code{\link[=TestPhiHurwitz]{TestPhiHurwitz()}},
\code{\link[=TestStability]{TestStability()}}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Simulation of State Space Models Data Functions}
\keyword{simStateSpace}
\keyword{ssm}
\keyword{test}
