% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simStateSpace-methods-statespacepb.R
\name{vcov.pbstatespace}
\alias{vcov.pbstatespace}
\title{Sampling Variance-Covariance Matrix Method for an Object of Class
\code{pbstatespace}}
\usage{
\method{vcov}{pbstatespace}(object, ...)
}
\arguments{
\item{object}{Object of Class \code{pbstatespace}.}

\item{...}{additional arguments.}
}
\value{
Returns the variance-covariance matrix of estimates.
}
\description{
Sampling Variance-Covariance Matrix Method for an Object of Class
\code{pbstatespace}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
