% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simStateSpace-methods-statespacepb.R
\name{vcov.statespacepb}
\alias{vcov.statespacepb}
\title{Sampling Variance-Covariance Matrix Method for an Object of Class
\code{statespacepb}}
\usage{
\method{vcov}{statespacepb}(object, ...)
}
\arguments{
\item{object}{Object of Class \code{statespacepb}.}

\item{...}{additional arguments.}
}
\value{
Returns the variance-covariance matrix of estimates.
}
\description{
Sampling Variance-Covariance Matrix Method for an Object of Class
\code{statespacepb}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
