% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simStateSpace-lin-sde-cov-y.R
\name{LinSDECovY}
\alias{LinSDECovY}
\title{Steady-State Covariance Matrix for the
Observed Variables in the
Linear Stochastic Differential Equation Model}
\usage{
LinSDECovY(lambda, theta, cov_eta)
}
\arguments{
\item{lambda}{Numeric matrix.
Factor loading matrix linking the latent variables
to the observed variables
(\eqn{\boldsymbol{\Lambda}}).}

\item{theta}{Numeric matrix.
The covariance matrix of
the measurement error
(\eqn{\boldsymbol{\Theta}}).}

\item{cov_eta}{Numeric matrix.
The steady-state covariance matrix
for the latent variables
in the linear stochastic differential equation model}
}
\description{
The steady-state covariance matrix
for the observed variables
in the linear stochastic differential equation model
\eqn{\mathrm{Cov} \left( \mathbf{y} \right)}
is given by
\deqn{
  \mathrm{Cov} \left( \mathbf{y} \right)
  =
  \boldsymbol{\Lambda}
  \mathrm{Cov} \left( \boldsymbol{\eta} \right)
  \boldsymbol{\Lambda}^{\prime}
  +
  \boldsymbol{\Theta}
}
where
\eqn{\boldsymbol{\Lambda}}
is the matrix of factor loadings,
\eqn{\boldsymbol{\Theta}}
is the covariance matrix of
the measurement error,
and
\eqn{\mathrm{Cov} \left( \boldsymbol{\eta} \right)}
is the steady-state covariance matrix
for the latent variables.
}
\examples{
phi <- matrix(
  data = c(
    -0.10, 0.05,
    0.05, -0.10
  ),
  nrow = 2
)
sigma <- matrix(
  data = c(
    2.79, 0.06,
    0.06, 3.27
  ),
  nrow = 2
)
lambda <- diag(2)
theta <- 0.01 * diag(2)
cov_eta <- LinSDECovEta(
  phi = phi,
  sigma = sigma
)
LinSDECovY(
  lambda = lambda,
  theta = theta,
  cov_eta = cov_eta
)

}
\seealso{
Other Simulation of State Space Models Data Functions: 
\code{\link{LinSDE2SSM}()},
\code{\link{LinSDECovEta}()},
\code{\link{LinSDEMeanEta}()},
\code{\link{LinSDEMeanY}()},
\code{\link{SSMCovEta}()},
\code{\link{SSMCovY}()},
\code{\link{SSMMeanEta}()},
\code{\link{SSMMeanY}()},
\code{\link{SimAlphaN}()},
\code{\link{SimBetaN}()},
\code{\link{SimCovDiagN}()},
\code{\link{SimCovN}()},
\code{\link{SimIotaN}()},
\code{\link{SimPhiN}()},
\code{\link{SimSSMFixed}()},
\code{\link{SimSSMIVary}()},
\code{\link{SimSSMLinGrowth}()},
\code{\link{SimSSMLinGrowthIVary}()},
\code{\link{SimSSMLinSDEFixed}()},
\code{\link{SimSSMLinSDEIVary}()},
\code{\link{SimSSMOUFixed}()},
\code{\link{SimSSMOUIVary}()},
\code{\link{SimSSMVARFixed}()},
\code{\link{SimSSMVARIVary}()},
\code{\link{TestPhi}()},
\code{\link{TestStability}()},
\code{\link{TestStationarity}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Simulation of State Space Models Data Functions}
\keyword{linsde}
\keyword{simStateSpace}
