% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simStateSpace-lin-sde-mean-y.R
\name{LinSDEMeanY}
\alias{LinSDEMeanY}
\title{Steady-State Mean Vector for the
Observed Variables in the
Linear Stochastic Differential Equation Model}
\usage{
LinSDEMeanY(nu, lambda, mean_eta)
}
\arguments{
\item{nu}{Numeric vector.
Vector of intercept values for the measurement model
(\eqn{\boldsymbol{\nu}}).}

\item{lambda}{Numeric matrix.
Factor loading matrix linking the latent variables
to the observed variables
(\eqn{\boldsymbol{\Lambda}}).}

\item{mean_eta}{Numeric vector.
Steady-state mean vector
of the latent variables
\eqn{\mathrm{Mean} \left( \boldsymbol{\eta} \right)}.}
}
\description{
The steady-state mean vector
for the observed variables
in the linear stochastic differential equation model
\eqn{\mathrm{Mean} \left( \boldsymbol{\eta} \right)}
is given by
\deqn{
  \boldsymbol{\nu}
  +
  \boldsymbol{\Lambda}
  \mathrm{Mean} \left( \boldsymbol{\eta} \right)
}
where
\eqn{\boldsymbol{\nu}}
is the vector of intercept values
for the measurement model,
\eqn{\boldsymbol{\Lambda}}
is the matrix of factor loadings,
and
\eqn{\mathrm{Mean} \left( \boldsymbol{\eta} \right)}
is the steady-state mean vector
for the latent variables.
}
\examples{
iota <- c(0.317, 0.230)
phi <- matrix(
  data = c(
    -0.10, 0.05,
    0.05, -0.10
  ),
  nrow = 2
)
nu <- rep(x = 0.03, times = 2)
lambda <- diag(2)
mean_eta <- LinSDEMeanEta(
  phi = phi,
  iota = iota
)
LinSDEMeanY(
  nu = nu,
  lambda = lambda,
  mean_eta = mean_eta
)

}
\seealso{
Other Simulation of State Space Models Data Functions: 
\code{\link{LinSDE2SSM}()},
\code{\link{LinSDECovEta}()},
\code{\link{LinSDECovY}()},
\code{\link{LinSDEMeanEta}()},
\code{\link{SSMCovEta}()},
\code{\link{SSMCovY}()},
\code{\link{SSMMeanEta}()},
\code{\link{SSMMeanY}()},
\code{\link{SimAlphaN}()},
\code{\link{SimBetaN}()},
\code{\link{SimCovDiagN}()},
\code{\link{SimCovN}()},
\code{\link{SimIotaN}()},
\code{\link{SimPhiN}()},
\code{\link{SimSSMFixed}()},
\code{\link{SimSSMIVary}()},
\code{\link{SimSSMLinGrowth}()},
\code{\link{SimSSMLinGrowthIVary}()},
\code{\link{SimSSMLinSDEFixed}()},
\code{\link{SimSSMLinSDEIVary}()},
\code{\link{SimSSMOUFixed}()},
\code{\link{SimSSMOUIVary}()},
\code{\link{SimSSMVARFixed}()},
\code{\link{SimSSMVARIVary}()},
\code{\link{TestPhi}()},
\code{\link{TestStability}()},
\code{\link{TestStationarity}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Simulation of State Space Models Data Functions}
\keyword{linsde}
\keyword{simStateSpace}
