% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simStateSpace-list-2-matrix.R
\name{Sim2Matrix}
\alias{Sim2Matrix}
\title{Simulation Output to Matrix}
\usage{
Sim2Matrix(x, eta = FALSE, long = TRUE)
}
\arguments{
\item{x}{R object.
Output of
\code{\link[=SimSSM]{SimSSM()}},
\code{\link[=SimSSMOU]{SimSSMOU()}},
\code{\link[=SimSSMVAR]{SimSSMVAR()}},
\code{\link[=SimSSMFixed]{SimSSMFixed()}},
\code{\link[=SimSSMOUFixed]{SimSSMOUFixed()}},
\code{\link[=SimSSMVARFixed]{SimSSMVARFixed()}},
\code{\link[=SimSSMVary]{SimSSMVary()}},
\code{\link[=SimSSMOUVary]{SimSSMOUVary()}}, or
\code{\link[=SimSSMVARVary]{SimSSMVARVary()}}.}

\item{eta}{Logical.
If \code{eta = TRUE}, include \code{eta}.
If \code{eta = FALSE}, exclude \code{eta}.}

\item{long}{Logical.
If \code{long = TRUE}, use long format.
If \code{long = FALSE}, use wide format.}
}
\value{
Returns a matrix of simulated data.
}
\description{
This function converts the output of
\code{\link[=SimSSM]{SimSSM()}},
\code{\link[=SimSSMOU]{SimSSMOU()}},
\code{\link[=SimSSMVAR]{SimSSMVAR()}},
\code{\link[=SimSSMFixed]{SimSSMFixed()}},
\code{\link[=SimSSMOUFixed]{SimSSMOUFixed()}},
\code{\link[=SimSSMVARFixed]{SimSSMVARFixed()}},
\code{\link[=SimSSMVary]{SimSSMVary()}},
\code{\link[=SimSSMOUVary]{SimSSMOUVary()}}, or
\code{\link[=SimSSMVARVary]{SimSSMVARVary()}}
to a matrix.
}
\examples{
# prepare parameters
set.seed(42)
k <- p <- 3
iden <- diag(k)
iden_sqrt <- chol(iden)
null_vec <- rep(x = 0, times = k)
n <- 5
mu0 <- null_vec
sigma0_sqrt <- iden_sqrt
alpha <- null_vec
beta <- diag(x = 0.50, nrow = k)
psi_sqrt <- iden_sqrt
nu <- null_vec
lambda <- iden
theta_sqrt <- chol(diag(x = 0.50, nrow = k))
time <- 50
burn_in <- 0

# generate data
ssm <- SimSSM(
  mu0 = mu0,
  sigma0_sqrt = sigma0_sqrt,
  alpha = alpha,
  beta = beta,
  psi_sqrt = psi_sqrt,
  nu = nu,
  lambda = lambda,
  theta_sqrt = theta_sqrt,
  type = 0,
  time = time,
  burn_in = burn_in
)

# list to matrix
mat <- Sim2Matrix(ssm, long = TRUE)
str(mat)
head(mat)
mat <- Sim2Matrix(ssm, long = FALSE)
str(mat)
head(mat)

# generate data
ssm <- SimSSMFixed(
  n = n,
  mu0 = mu0,
  sigma0_sqrt = sigma0_sqrt,
  alpha = alpha,
  beta = beta,
  psi_sqrt = psi_sqrt,
  nu = nu,
  lambda = lambda,
  theta_sqrt = theta_sqrt,
  type = 0,
  time = time,
  burn_in = burn_in
)

# list to matrix
mat <- Sim2Matrix(ssm, long = TRUE)
str(mat)
head(mat)
mat <- Sim2Matrix(ssm, long = FALSE)
str(mat)
head(mat)

}
\seealso{
Other Simulation of State Space Models Data Functions: 
\code{\link{OU2SSM}()},
\code{\link{SimSSMFixed}()},
\code{\link{SimSSMLinGrowth}()},
\code{\link{SimSSMOUFixed}()},
\code{\link{SimSSMOUVary}()},
\code{\link{SimSSMOU}()},
\code{\link{SimSSMVARFixed}()},
\code{\link{SimSSMVARVary}()},
\code{\link{SimSSMVAR}()},
\code{\link{SimSSMVary}()},
\code{\link{SimSSM}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Simulation of State Space Models Data Functions}
\keyword{misc}
\keyword{simStateSpace}
