% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SimCovDiagN}
\alias{SimCovDiagN}
\title{Simulate Diagonal Covariance Matrices
from the Multivariate Normal Distribution}
\usage{
SimCovDiagN(n, sigma_diag, vcov_sigma_diag_l)
}
\arguments{
\item{n}{Positive integer.
Number of replications.}

\item{sigma_diag}{Numeric matrix.
The covariance matrix
(\eqn{\boldsymbol{\Sigma}}).}

\item{vcov_sigma_diag_l}{Numeric matrix.
Cholesky factorization (\code{t(chol(vcov_sigma_vech))})
of the sampling variance-covariance matrix of
\eqn{\mathrm{vech} \left( \boldsymbol{\Sigma} \right)}.}
}
\value{
Returns a list of random diagonal covariance matrices.
}
\description{
This function simulates random diagonal covariance matrices
from the multivariate normal distribution.
The function ensures that the generated covariance matrices
are positive semi-definite.
}
\examples{
n <- 10
sigma_diag <- c(1, 1, 1)
vcov_sigma_diag_l <- t(chol(0.001 * diag(3)))
SimCovDiagN(
  n = n,
  sigma_diag = sigma_diag,
  vcov_sigma_diag_l = vcov_sigma_diag_l
)

}
\seealso{
Other Simulation of State Space Models Data Functions: 
\code{\link{LinSDE2SSM}()},
\code{\link{LinSDECovEta}()},
\code{\link{LinSDECovY}()},
\code{\link{LinSDEMeanEta}()},
\code{\link{LinSDEMeanY}()},
\code{\link{ProjectToStability}()},
\code{\link{SSMCovEta}()},
\code{\link{SSMCovY}()},
\code{\link{SSMMeanEta}()},
\code{\link{SSMMeanY}()},
\code{\link{SimAlphaN}()},
\code{\link{SimBetaN}()},
\code{\link{SimBetaN2}()},
\code{\link{SimCovN}()},
\code{\link{SimIotaN}()},
\code{\link{SimPhiN}()},
\code{\link{SimSSMFixed}()},
\code{\link{SimSSMIVary}()},
\code{\link{SimSSMLinGrowth}()},
\code{\link{SimSSMLinGrowthIVary}()},
\code{\link{SimSSMLinSDEFixed}()},
\code{\link{SimSSMLinSDEIVary}()},
\code{\link{SimSSMOUFixed}()},
\code{\link{SimSSMOUIVary}()},
\code{\link{SimSSMVARFixed}()},
\code{\link{SimSSMVARIVary}()},
\code{\link{SpectralRadius}()},
\code{\link{TestPhi}()},
\code{\link{TestStability}()},
\code{\link{TestStationarity}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Simulation of State Space Models Data Functions}
\keyword{simStateSpace}
\keyword{ssm}
