% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simStateSpace-sim-mvn.R
\name{SimMuN}
\alias{SimMuN}
\title{Simulate Set Point Vectors
from the Multivariate Normal Distribution}
\usage{
SimMuN(n, mu, vcov_mu_l)
}
\arguments{
\item{n}{Positive integer.
Number of replications.}

\item{mu}{Numeric vector.
Set point (\eqn{\boldsymbol{\mu}}).}

\item{vcov_mu_l}{Numeric matrix.
Cholesky factorization (\code{t(chol(vcov_mu))})
of the sampling variance-covariance matrix of
\eqn{\boldsymbol{\mu}}.}
}
\value{
Returns a list of random set point vectors.
}
\description{
This function simulates random set point vectors
from the multivariate normal distribution.
}
\examples{
n <- 10
mu <- c(0, 0, 0)
vcov_mu_l <- t(chol(0.001 * diag(3)))
SimMuN(n = n, mu = mu, vcov_mu_l = vcov_mu_l)

}
\seealso{
Other Simulation of State Space Models Data Functions:
\code{\link{LinSDE2SSM}()},
\code{\link{LinSDECovEta}()},
\code{\link{LinSDECovY}()},
\code{\link{LinSDEInterceptEta}()},
\code{\link{LinSDEInterceptY}()},
\code{\link{LinSDEMeanEta}()},
\code{\link{LinSDEMeanY}()},
\code{\link{ProjectToHurwitz}()},
\code{\link{ProjectToStability}()},
\code{\link{SSMCovEta}()},
\code{\link{SSMCovY}()},
\code{\link{SSMInterceptEta}()},
\code{\link{SSMInterceptY}()},
\code{\link{SSMMeanEta}()},
\code{\link{SSMMeanY}()},
\code{\link{SimAlphaN}()},
\code{\link{SimBetaN}()},
\code{\link{SimBetaN2}()},
\code{\link{SimBetaNCovariate}()},
\code{\link{SimCovDiagN}()},
\code{\link{SimCovN}()},
\code{\link{SimIotaN}()},
\code{\link{SimMVN}()},
\code{\link{SimNuN}()},
\code{\link{SimPhiN}()},
\code{\link{SimPhiN2}()},
\code{\link{SimPhiNCovariate}()},
\code{\link{SimSSMFixed}()},
\code{\link{SimSSMIVary}()},
\code{\link{SimSSMLinGrowth}()},
\code{\link{SimSSMLinGrowthIVary}()},
\code{\link{SimSSMLinSDEFixed}()},
\code{\link{SimSSMLinSDEIVary}()},
\code{\link{SimSSMOUFixed}()},
\code{\link{SimSSMOUIVary}()},
\code{\link{SimSSMVARFixed}()},
\code{\link{SimSSMVARIVary}()},
\code{\link{SpectralRadius}()},
\code{\link{TestPhi}()},
\code{\link{TestPhiHurwitz}()},
\code{\link{TestStability}()},
\code{\link{TestStationarity}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Simulation of State Space Models Data Functions}
\keyword{simStateSpace}
\keyword{ssm}
