% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simStateSpace-package.R
\docType{package}
\name{simStateSpace-package}
\alias{simStateSpace}
\alias{simStateSpace-package}
\title{simStateSpace: Simulate Data from State Space Models}
\description{
Provides a streamlined and user-friendly framework for simulating data in state space models, particularly when the number of subjects/units (n) exceeds one, a scenario commonly encountered in social and behavioral sciences. For an introduction to state space models in social and behavioral sciences, refer to Chow, Ho, Hamaker, and Dolan (2010) \doi{10.1080/10705511003661553}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jeksterslab/simStateSpace}
  \item \url{https://jeksterslab.github.io/simStateSpace/}
  \item Report bugs at \url{https://github.com/jeksterslab/simStateSpace/issues}
}

}
\author{
\strong{Maintainer}: Ivan Jacob Agaloos Pesigan \email{r.jeksterslab@gmail.com} (\href{https://orcid.org/0000-0003-4818-8420}{ORCID}) [copyright holder]

}
\keyword{internal}
