% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SSMMeanEta}
\alias{SSMMeanEta}
\title{State Mean Vector for the
State Space Model}
\usage{
SSMMeanEta(beta, alpha)
}
\arguments{
\item{beta}{Numeric matrix.
The transition matrix (\eqn{\boldsymbol{\beta}}).}

\item{alpha}{Numeric vector.
Vector of constant values for the dynamic model
(\eqn{\boldsymbol{\alpha}}).}
}
\description{
This function calculates the state mean vector
for the state space model
given by
\deqn{
  \mathrm{Mean} \left( \boldsymbol{\eta} \right)
  =
  \left(
    \mathbf{I} - \boldsymbol{\beta}
  \right)^{-1}
  \boldsymbol{\alpha} .
}
}
\examples{
beta <- 0.50 * diag(3)
alpha <- rep(x = 0.001, times = 3)
SSMMeanEta(beta = beta, alpha = alpha)

}
\seealso{
Other Simulation of State Space Models Data Functions: 
\code{\link{LinSDE2SSM}()},
\code{\link{LinSDECov}()},
\code{\link{LinSDEMean}()},
\code{\link{SSMCovEta}()},
\code{\link{SSMCovY}()},
\code{\link{SSMMeanY}()},
\code{\link{SimAlphaN}()},
\code{\link{SimBetaN}()},
\code{\link{SimCovDiagN}()},
\code{\link{SimCovN}()},
\code{\link{SimIotaN}()},
\code{\link{SimPhiN}()},
\code{\link{SimSSMFixed}()},
\code{\link{SimSSMIVary}()},
\code{\link{SimSSMLinGrowth}()},
\code{\link{SimSSMLinGrowthIVary}()},
\code{\link{SimSSMLinSDEFixed}()},
\code{\link{SimSSMLinSDEIVary}()},
\code{\link{SimSSMOUFixed}()},
\code{\link{SimSSMOUIVary}()},
\code{\link{SimSSMVARFixed}()},
\code{\link{SimSSMVARIVary}()},
\code{\link{TestPhi}()},
\code{\link{TestStability}()},
\code{\link{TestStationarity}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Simulation of State Space Models Data Functions}
\keyword{simStateSpace}
\keyword{ssm}
