% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SSMMeanY}
\alias{SSMMeanY}
\title{Observed Variable Mean Vector for the
State Space Model}
\usage{
SSMMeanY(nu, lambda, mean_eta)
}
\arguments{
\item{nu}{Numeric vector.
Vector of constant values for the measurement model
(\eqn{\boldsymbol{\nu}}).}

\item{lambda}{Numeric matrix.
The factor loadings matrix (\eqn{\boldsymbol{\Lambda}}).}

\item{mean_eta}{Numeric vector.
State mean vector
\eqn{\mathrm{Mean} \left( \boldsymbol{\eta} \right)}.}
}
\description{
This function calculates the observed variable mean vector
for the state space model
given by
\deqn{
  \mathrm{Mean} \left( \mathbf{y} \right)
  =
  \boldsymbol{\nu}
  +
  \boldsymbol{\Lambda}
  \mathrm{Mean} \left( \boldsymbol{\eta} \right) .
}
}
\examples{
beta <- 0.50 * diag(3)
alpha <- rep(x = 0.001, times = 3)
nu <- rep(x = 0.03, times = 3)
lambda <- diag(3)
mean_eta <- SSMMeanEta(beta = beta, alpha = alpha)
SSMMeanY(
  nu = nu,
  lambda = lambda,
  mean_eta = mean_eta
)

}
\seealso{
Other Simulation of State Space Models Data Functions: 
\code{\link{LinSDE2SSM}()},
\code{\link{LinSDECov}()},
\code{\link{LinSDEMean}()},
\code{\link{SSMCovEta}()},
\code{\link{SSMCovY}()},
\code{\link{SSMMeanEta}()},
\code{\link{SimAlphaN}()},
\code{\link{SimBetaN}()},
\code{\link{SimCovDiagN}()},
\code{\link{SimCovN}()},
\code{\link{SimIotaN}()},
\code{\link{SimPhiN}()},
\code{\link{SimSSMFixed}()},
\code{\link{SimSSMIVary}()},
\code{\link{SimSSMLinGrowth}()},
\code{\link{SimSSMLinGrowthIVary}()},
\code{\link{SimSSMLinSDEFixed}()},
\code{\link{SimSSMLinSDEIVary}()},
\code{\link{SimSSMOUFixed}()},
\code{\link{SimSSMOUIVary}()},
\code{\link{SimSSMVARFixed}()},
\code{\link{SimSSMVARIVary}()},
\code{\link{TestPhi}()},
\code{\link{TestStability}()},
\code{\link{TestStationarity}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Simulation of State Space Models Data Functions}
\keyword{simStateSpace}
\keyword{ssm}
